omap: serial: Fix bug when transmitting more than 16 bytes
Current driver used an inverted polarity to assert the RTS line.
It worked only because the FIFO empty interrupt level is 16 bytes.
So basically the empty interrupt was the one asserting the line and
that is why you could transmit up to 16 bytes.

This patch inverts the polarity and also removes the RTS update in
serial_omap_stop_tx() since the RTS line has to remain asserted until
the serial_omap_irq() is executed due an empty interrupt raised and
the condition __serial_omap_tx_empty() is true.

Signed-off-by: Javier Martinez Canillas <javier@dowhile0.org>