Skip a frame buffer update on a dedicated kernel parameter.
This commit introduces a "omapfb.skipupd" kernel parameter
which skips invocation of panel's update() during omapfb init.

The rationale:
We have a uboot that has already displayed an image and
we want that image to last all the way through kernel boot.
Our LCD is a DSI COMMAND mode LCD, and though opamfb driver
zeros out the framebuffer memory, the LCD will not turn
blank until the update. And it's the update we prevent.
Until the update the LCD keeps showing  u-boot's image.